Modularize CLI (#59)
* Move CLI utils to its own module

* Move CLI main file to its own module

* Add __init__.py file

* Change CLI entrypoint in the pyproject.toml file
